move margin top value to consts
