+++ src/components/prayers/ComplineBlessingAndConcludingVersicle.tsx
@@ -1,4 +1,5 @@
 import { V, R } from './index';
+import { styles } from 'consts/styles';
 
 type PropsType = Readonly<{
   variant: 1 | 2;
@@ -8,7 +9,7 @@ export const ComplineBlessingAndConcludingVersicle = ({
   variant,
 }: PropsType) => {
   return (
-    <div className="mt-6">
+    <div className={styles.marginTop}>
       <V>
         {variant === 1
           ? 'Noc spokojną i śmierć szczęśliwą niech nam da Bóg wszechmogący, Ojciec i Syn, i Duch Święty.'
