+++ src/components/prayers/KyrieEleison.tsx
@@ -1,8 +1,9 @@
 import { V, R } from './index';
+import { styles } from 'consts/styles';
 
 export const KyrieEleison = () => {
   return (
-    <div className="mt-6">
+    <div className={styles.marginTop}>
       <V>Kyrie eleison.</V>
       <R>Chryste eleison, Kyrie eleison.</R>
     </div>
