+++ src/consts/styles.ts
@@ -1,0 +1,3 @@
+export const styles = {
+  marginTop: 'mt-6',
+};
